+++ ft_printf/printffone/itoa_base_ptr.c
@@ -12,9 +12,9 @@
 
 #include "ft_printf.h"
 
-static unsigned int	getpower(unsigned int count, int basenum)
+static long int	getpower(unsigned int count, int basenum)
 {
-	unsigned int	i;
+	long int	i;
 
 	i = 1;
 	while (count > 1)
@@ -26,10 +26,10 @@ static unsigned int	getpower(unsigned int count, int basenum)
 }
 
 
-char	*ft_itoa_base_ptr(uintptr_t number, char *base)
+char	*ft_itoa_base_ptr(long int number, char *base)
 {
 	int basen;
-	int ipwr;
+	long int ipwr;
 	char *ret;
 	int i;
